requirements.txt
django==2.1.7
djangorestframework==3.9.2
dj-database-url==0.5.0
psycopg2==2.7.7
pillow==5.4.1
django-storages==1.7.1
boto3==1.9.120
django-cors-middleware==1.3.1
gunicorn==19.9.0
django-solo==1.1.3
